rename interactive-e2e to interactive
