+++ packages/server/test/integration/cypress_spec.js
@@ -22,7 +22,7 @@ const ciProvider = require(`../../lib/util/ci_provider`)
 const settings = require(`../../lib/util/settings`)
 const Events = require(`../../lib/gui/events`)
 const Windows = require(`../../lib/gui/windows`)
-const interactiveMode = require(`../../lib/modes/interactive-e2e`)
+const interactiveMode = require(`../../lib/modes/interactive`)
 const runMode = require(`../../lib/modes/run`)
 const api = require(`../../lib/api`)
 const cwd = require(`../../lib/cwd`)
